feat: add new Generate facade and Generate class for MadeForYou\Helpers package
The `registeringPackage` method in the `PackageServiceProvider` class has been updated to include the binding for the `Generate` facade and the `packages` binding has been updated to use a closure. This change allows for easier access to the `Generate` class and the `packages` instance.

A new `Generate` facade and `Generate` class have been added to the `MadeForYou\Helpers` namespace. The `Generate` class provides a `filamentLink` method that generates an HTML link with specific attributes for use with the Filament package.

These changes improve the functionality and flexibility of the `MadeForYou\Helpers` package.