chore: add filamentLink method to Generate facade
The `Generate` facade in the `MadeForYou\Helpers\Facades` namespace has been updated to include a new `filamentLink` method. This method allows for generating an HTML link with a specified link, title, and optional tooltip. The method returns an `HtmlString` object.

Additionally, the aliases in the `composer.json` file have been updated to include the `Generate` facade.

This change was made to provide a convenient way to generate HTML links with custom attributes using the `Generate` facade.